requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.8.2
cachelib==0.13.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-WTF==1.2.1
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
psycopg2-binary==2.9.9
pycparser==2.22
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
